Update ReadGrib
I have replaced looping through seperate grib2 files, with vectorization. This results in the faster execution of the function especially when many grib2 files are to be read. 
There is no change in the function arguments or outputs, users can apply ReadGrib function like before.
There is no change in reading grib1 files